backend/requirements.txt
asgiref==3.8.1
astroid==2.14.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dill==0.3.6
Django==5.0.7
django-cors-headers==4.4.0
django-debug-toolbar==4.2.0
django-mailbox==4.8.2
django-shopping-cart==0.1
djangorestframework==3.15.2
Flask==2.2.3
Flask-MySQLdb==1.0.1
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.1.1
Pillow==9.4.0
platformdirs==3.0.0
protobuf==4.21.12
psycopg2==2.9.7
pylint==2.16.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2023.3.post1
razorpay==1.3.0
requests==2.28.2
shell==1.0.1
six==1.16.0
sqlparse==0.4.3
tomlkit==0.11.6
tzdata==2022.7
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.14.1
"gunicorn" 
